Create Auxiliary Script
Start working on auxiliary functions that aren't used in the automated processing. The first trim_null_bytes can trim null bytes and preserve extra data in the overlay if there is any.

This is the first of scripts, but I need to generalize this function to work for bytes other than null. That being said, there are times this can accomplish things that the automatic processing currently cannot.