Fix alignment bug + trim improvements
Previous version was subtracting instead of adding the alignment. This has been fixed in this version.

The find_chunk_start method had some logic which was kind of weird. That has been clarified and the trimming method will always leave at the least one or two portions so as to not remove too much.